refactor: 게시물 옵션 모달 표시
