+++ src/components/Home/OurTownPost/PostComponents/PostComponent.tsx
@@ -1,9 +1,21 @@
+import styled from 'styled-components';
 import ImgSection from './ImgSection';
 import ContentSection from './ContentSection';
 import SebSection from './SebSection';
+import useModal from '@hooks/useModal';
+import BottomSheet from '@components/common/BottomSheet';
+import ConfirmModal from '@components/common/ConfirmModal';
 
+const ColorBlackMenu = styled.div`
+  color: ${({ theme }) => theme.color.black};
+`;
+
+const DeletePost = styled.div`
+  color: ${({ theme }) => theme.color.orange};
+`;
 interface IPostComponentProps {
   post: {
+    id: string;
     type: number;
     image: string;
     title: string;
@@ -14,14 +26,27 @@ interface IPostComponentProps {
     date: string;
   };
   isOption?: boolean;
-  onOptionClick?: () => void;
 }
 
-const PostComponent = ({
-  post,
-  isOption = false,
-  onOptionClick = () => {},
-}: IPostComponentProps) => {
+const PostComponent = ({ post, isOption = false }: IPostComponentProps) => {
+  const { isOpen, open, close } = useModal();
+  const {
+    isOpen: isDeleteModalOpen,
+    open: deleteModalOpen,
+    close: deleteModalClose,
+  } = useModal();
+
+  const handleEditPost = () => {
+    console.log(post.id);
+  };
+  const handleChangeStatus = () => {
+    console.log(post.id);
+  };
+  const handleDeletePost = () => {
+    console.log(post.id);
+    deleteModalClose();
+  };
+
   return (
     <>
       <ImgSection type={post.type} image={post.image} />
@@ -35,7 +60,31 @@ const PostComponent = ({
         location={post.location}
         date={post.date}
         isOption={isOption}
-        onOptionClick={onOptionClick}
+        onOptionClick={open}
+      />
+      {isOpen && (
+        <BottomSheet height="250px" onClick={close}>
+          <ColorBlackMenu onClick={handleEditPost}>게시물 수정</ColorBlackMenu>
+          <ColorBlackMenu onClick={handleChangeStatus}>
+            상태 변경
+          </ColorBlackMenu>
+          <DeletePost
+            onClick={() => {
+              close();
+              deleteModalOpen();
+            }}
+          >
+            삭제
+          </DeletePost>
+        </BottomSheet>
+      )}
+      <ConfirmModal
+        isOpen={isDeleteModalOpen}
+        title="게시물 삭제"
+        content="게시물을 삭제하시겠습니까?"
+        confirmedContent="게시물이 삭제되었습니다."
+        onFinalOkClick={handleDeletePost}
+        closeAction={deleteModalClose}
       />
     </>
   );
